Download the latest version of the openapi generate client
