+++ runtime/src/main/resources/application.properties
@@ -30,7 +30,8 @@ terraform.debug.level=DEBUG
 opentofu.workspace.directory=xpanse_deploy_opentofu_ws
 opentofu.debug.enabled=false
 opentofu.debug.level=DEBUG
-openapi.download-generator-client-url=https://repo1.maven.org/maven2/org/openapitools/openapi-generator-cli/6.6.0/openapi-generator-cli-6.6.0.jar
+openapi.generator.client.version=@openapi.generator.plugin.version@
+openapi.generator.client.download-url=https://repo1.maven.org/maven2/org/openapitools/openapi-generator-cli/${openapi.generator.client.version}/openapi-generator-cli-${openapi.generator.client.version}.jar
 openapi.path=openapi/
 openapi.url=/openapi/*
 http.logging.enabled=true
