+++ runtime/src/test/java/org/eclipse/xpanse/runtime/modules/common/openapi/OpenApiCommonTest.java
@@ -1,9 +1,4 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- * SPDX-FileCopyrightText: Huawei Inc.
- */
-
-package org.eclipse.xpanse.runtime.modules.common;
+package org.eclipse.xpanse.runtime.modules.common.openapi;
 
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
@@ -12,11 +7,15 @@ import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.URI;
 import java.net.URL;
+import java.util.UUID;
 import org.eclipse.xpanse.common.openapi.OpenApiGeneratorJarManage;
 import org.eclipse.xpanse.common.openapi.OpenApiUrlManage;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.MethodOrderer;
+import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -24,18 +23,29 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+
 @ContextConfiguration(classes = {OpenApiUrlManage.class, OpenApiGeneratorJarManage.class})
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(properties = {"spring.profiles.active=default"})
-class OpenApiGeneratorJarManageTest {
-
-    @Autowired
-    private OpenApiGeneratorJarManage openApiGeneratorJarManage;
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+class OpenApiCommonTest {
 
     @Value("${openapi.path}")
     private String openApiPath;
-    @Value("${openapi.download-generator-client-url}")
+    @Value("${openapi.generator.client.download-url}")
     private String clientDownloadUrl;
+    @Value("${server.port}")
+    private Integer serverPort;
+    @Autowired
+    private OpenApiGeneratorJarManage openApiGeneratorJarManage;
+    @Autowired
+    private OpenApiUrlManage openApiUrlManage;
+
+    @BeforeEach
+    void setUp() {
+        openApiUrlManage = new OpenApiUrlManage(openApiPath, serverPort);
+        openApiGeneratorJarManage = new OpenApiGeneratorJarManage(clientDownloadUrl, openApiPath);
+    }
 
     @Test
     void testGetOpenApiWorkdir() {
@@ -44,7 +54,23 @@ class OpenApiGeneratorJarManageTest {
     }
 
     @Test
-    @Disabled
+    void testGetClientDownLoadUrl() {
+        // Run the test
+        String result = openApiGeneratorJarManage.getClientDownLoadUrl();
+        // Verify the results
+        Assertions.assertEquals(clientDownloadUrl, result);
+    }
+
+    @Test
+    void testGetOpenapiPath() {
+        // Run the test
+        String result = openApiGeneratorJarManage.getOpenapiPath();
+        // Verify the results
+        Assertions.assertEquals(openApiPath, result);
+    }
+
+    @Test
+    @Order(1)
     void testDownloadClientJar() throws IOException {
         // SetUp
         File jarFile = openApiGeneratorJarManage.getCliFile();
@@ -67,19 +93,24 @@ class OpenApiGeneratorJarManageTest {
     }
 
     @Test
-    void testGetClientDownLoadUrl() {
+    @Order(2)
+    void testGetServiceUrl() {
+        // SetUp
         // Run the test
-        String result = openApiGeneratorJarManage.getClientDownLoadUrl();
+        String result = openApiUrlManage.getServiceUrl();
         // Verify the results
-        Assertions.assertEquals(clientDownloadUrl, result);
+        Assertions.assertEquals("http://localhost", result);
     }
 
     @Test
-    void testGetOpenapiPath() {
+    void testGetOpenApiUrl() {
+        // SetUp
+        String id = UUID.randomUUID().toString();
+        String openApiUrl =
+                openApiUrlManage.getServiceUrl() + "/" + openApiPath + id + ".html";
         // Run the test
-        String result = openApiGeneratorJarManage.getOpenapiPath();
+        String result = openApiUrlManage.getOpenApiUrl(id);
         // Verify the results
-        Assertions.assertEquals(openApiPath, result);
+        Assertions.assertEquals(openApiUrl, result);
     }
 }
-
